requirements.txt
altair==6.0.0
attrs==25.4.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
choreographer==1.2.1
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
defusedxml==0.7.1
fonttools==4.62.1
fpdf==1.7.2
fpdf2==2.8.7
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
googleapis-common-protos==1.73.0
greenlet==3.3.2
httplib2==0.31.2
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kaleido==1.2.0
logistro==2.0.1
MarkupSafe==3.0.3
narwhals==2.18.0
numpy==2.4.3
oauthlib==3.3.1
orjson==3.11.7
packaging==26.0
pandas==2.3.3
pillow==12.1.1
playwright==1.58.0
plotly==6.6.0
pluggy==1.6.0
proto-plus==1.27.1
protobuf==6.33.5
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydeck==0.9.1
pyee==13.0.1
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
simplejson==3.20.2
six==1.17.0
smmap==5.0.3
streamlit==1.55.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
watchdog==6.0.0